+++ java/src/main/java/com/bazaarvoice/gumshoe/Configuration.java
@@ -1,46 +1,33 @@
 package com.bazaarvoice.gumshoe;
 
 /**
- * Configuration to control behavior of Gumshoe.  Allows specification of:
- *
- * - Filter used to filter out events emitted
- * - Decorator used to decorate all events with global data
- * - Publisher used to publish events in some fashion.
- *
+ * Interface of objects that can be used to configure Gumshoe.
+ * 
  * @author lance.woodson
  *
  */
-public class Configuration {
-    private Filter filter;
-    private Decorator decorator;
-    private Publisher publisher;
-
-    public Configuration() {
-        // TODO wire up default implementations of Filter, Decorator
-        // and Publisher
-    }
-
-    public Filter getFilter() {
-        return filter;
-    }
-
-    public void setFilter(Filter filter) {
-        this.filter = filter;
-    }
-
-    public Decorator getDecorator() {
-        return decorator;
-    }
-
-    public void setDecorator(Decorator decorator) {
-        this.decorator = decorator;
-    }
-
-    public Publisher getPublisher() {
-        return publisher;
-    }
-
-    public void setPublisher(Publisher publisher) {
-        this.publisher = publisher;
-    }
+public interface Configuration {
+    /**
+     * The name of the application
+     * 
+     * @return
+     */
+    String getApplicationName();
+    /**
+     * Object that decides whether an event will be published or not
+     * 
+     * @return
+     */
+    Filter getFilter();
+    /**
+     * Object that decorates all events with global information
+     * 
+     * @return
+     */
+    Decorator getDecorator();
+    /**
+     * Object that publishes the event in some way
+     * @return
+     */
+    Publisher getPublisher();
 }
